feat: inline shallowEqualOptions function into useLifecycle for better encapsulation
